+++ src/RTT_Format_Reader/Nodes.cc
@@ -69,12 +69,12 @@ void Nodes::readData(ifstream &meshfile) {
   string dummyString;
   int nodeNum;
 
-  for (size_t i = 0; i < static_cast<size_t>(dims.get_nnodes()); ++i) {
+  for (size_t i = 0; i < dims.get_nnodes(); ++i) {
     nodeNum = readNextInt(meshfile);
     // meshfile >> nodeNum;
     Insist(static_cast<size_t>(nodeNum) == i + 1, "Invalid mesh file: node index out of order");
     Check(i < coords.size());
-    for (size_t j = 0; j < static_cast<size_t>(dims.get_ndim()); ++j) {
+    for (size_t j = 0; j < dims.get_ndim(); ++j) {
       Check(j < coords[i].size());
       meshfile >> coords[i][j];
     }
@@ -82,7 +82,7 @@ void Nodes::readData(ifstream &meshfile) {
     meshfile >> parents[i];
     --parents[i];
     Check(i < flags.size());
-    for (size_t j = 0; j < static_cast<size_t>(dims.get_nnode_flag_types()); ++j) {
+    for (size_t j = 0; j < dims.get_nnode_flag_types(); ++j) {
       Check(j < flags[i].size());
       meshfile >> flags[i][j];
       Check(j < INT_MAX);
