+++ src/ds++/Release.cc
@@ -188,7 +188,7 @@ const std::string copyright() {
 //! This version can be called by Fortran and wraps the C++ version.
 extern "C" void ec_release(char *release_string, size_t maxlen) {
   std::string tmp_rel = rtt_dsxx::release();
-  if (tmp_rel.size() >= static_cast<size_t>(maxlen)) {
+  if (tmp_rel.size() >= maxlen) {
     tmp_rel = tmp_rel.substr(0, maxlen - 1);
   }
   std::memcpy(release_string, tmp_rel.c_str(), tmp_rel.size() + 1);
